+++ src/collective/querynextprev/tests/__init__.py
@@ -4,8 +4,13 @@ import json
 
 query = json.dumps({
     'portal_type': 'Document',
-    'sort_on': 'sortable_title',
-    'title': 'é'
+    'sort_on': 'sortable_title'
+})
+
+query_utf8 = json.dumps({
+    'portal_type': 'Document',
+    'sort_on': 'id',
+    'Title': 'é'
 })